Power Overwhelming!
Using TeamReborn's Energy API.

+++ core/src/main/java/io/github/shkschneider/awesome/core/AwesomeUtils.kt
@@ -27,7 +27,7 @@ object AwesomeUtils {
         )
     }
 
-    fun humanReadable(n: Int): String =
+    fun humanReadable(n: Long): String =
         when {
             n < 1_000 -> String.format("%d", n)
             n < 1_000_000 -> String.format("%.01fK", n.toFloat() / 1_000)
